+++ src/main/java/com/moratan251/psitweaks/client/ClientSetup.java
@@ -2,7 +2,6 @@ package com.moratan251.psitweaks.client;
 
 import com.moratan251.psitweaks.Psitweaks;
 import com.moratan251.psitweaks.client.gui.machine.GuiPortableCADAssembler;
-import com.moratan251.psitweaks.client.gui.machine.GuiProgramResearcher;
 import com.moratan251.psitweaks.client.gui.machine.ModMenuTypes;
 import com.moratan251.psitweaks.client.renderer.EntityTimeAcceleratorRenderer;
 import com.moratan251.psitweaks.client.renderer.FlareCircleRenderer;
@@ -39,7 +38,6 @@ public class ClientSetup {
     public static void onClientSetup(FMLClientSetupEvent event) {
         event.enqueueWork(() -> {
             MenuScreens.register(ModMenuTypes.PORTABLE_CAD_ASSEMBLER.get(), GuiPortableCADAssembler::new);
-            MenuScreens.register(ModMenuTypes.PROGRAM_RESEARCHER.get(), GuiProgramResearcher::new);
         });
     }
 }