+++ src/main/java/com/moratan251/psitweaks/common/registries/PsitweaksMekanismContainerTypes.java
@@ -2,6 +2,7 @@ package com.moratan251.psitweaks.common.registries;
 
 import com.moratan251.psitweaks.Psitweaks;
 import com.moratan251.psitweaks.common.tile.machine.TileEntityMaterialMutator;
+import com.moratan251.psitweaks.common.tile.machine.ProgramResearcherBlockEntity;
 import com.moratan251.psitweaks.common.tile.machine.TileEntityPsionicGenerator;
 import com.moratan251.psitweaks.common.tile.machine.TileEntitySculkEroder;
 import com.moratan251.psitweaks.common.tile.machine.TileEntityTranscendentEnergyCube;
@@ -16,6 +17,9 @@ public class PsitweaksMekanismContainerTypes {
 
     public static final ContainerTypeRegistryObject<MekanismTileContainer<TileEntitySculkEroder>> SCULK_ERODER =
             CONTAINER_TYPES.register(PsitweaksMekanismBlocks.SCULK_ERODER, TileEntitySculkEroder.class);
+
+    public static final ContainerTypeRegistryObject<MekanismTileContainer<ProgramResearcherBlockEntity>> PROGRAM_RESEARCHER =
+            CONTAINER_TYPES.register(PsitweaksMekanismBlocks.PROGRAM_RESEARCHER, ProgramResearcherBlockEntity.class);
 
     public static final ContainerTypeRegistryObject<MekanismTileContainer<TileEntityMaterialMutator>> MATERIAL_MUTATOR =
             CONTAINER_TYPES.register(PsitweaksMekanismBlocks.MATERIAL_MUTATOR, TileEntityMaterialMutator.class);